moneychanger_ai_app/requirements.txt
python-dotenv==1.1.1
requests==2.32.4
streamlit==1.47.0
openai==1.97.0
langsmith==0.4.6